# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Billing application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ── Customers ──
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report for period
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [daily, monthly, yearly]
            default: monthly
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/gst:
    get:
      operationId: getGstReport
      tags: [reports]
      summary: Get GST report for financial year
      parameters:
        - name: financialYear
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: GST report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstReport"
  /reports/profit:
    get:
      operationId: getProfitReport
      tags: [reports]
      summary: Get profit report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [daily, monthly, yearly]
            default: monthly
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
  /reports/product-wise:
    get:
      operationId: getProductWiseReport
      tags: [reports]
      summary: Product-wise sales report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product-wise report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWiseReport"
  /reports/customer-wise:
    get:
      operationId: getCustomerWiseReport
      tags: [reports]
      summary: Customer-wise sales report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer-wise report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerWiseReport"

  # ── Backup ──
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, costPrice, marginPercent, sellingPrice, gstPercent, stock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        costPrice:
          type: number
        marginPercent:
          type: number
        sellingPrice:
          type: number
        gstPercent:
          type: number
        stock:
          type: integer
        hsn:
          type: string
        unit:
          type: string
        createdAt:
          type: string
    CreateProductBody:
      type: object
      required: [name, costPrice, marginPercent, gstPercent, stock]
      properties:
        name:
          type: string
        costPrice:
          type: number
        marginPercent:
          type: number
        gstPercent:
          type: number
        stock:
          type: integer
        hsn:
          type: string
        unit:
          type: string
          default: "pcs"
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        costPrice:
          type: number
        marginPercent:
          type: number
        gstPercent:
          type: number
        stock:
          type: integer
        hsn:
          type: string
        unit:
          type: string
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    SalesReport:
      type: object
      required: [entries, totalSales, totalGst]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/SalesReportEntry"
        totalSales:
          type: number
        totalGst:
          type: number
        totalProfit:
          type: number
    SalesReportEntry:
      type: object
      required: [period, sales, gst, invoiceCount]
      properties:
        period:
          type: string
        sales:
          type: number
        gst:
          type: number
        profit:
          type: number
        invoiceCount:
          type: integer
    GstReport:
      type: object
      required: [entries, totalCgst, totalSgst, totalIgst, totalGst]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/GstReportEntry"
        totalCgst:
          type: number
        totalSgst:
          type: number
        totalIgst:
          type: number
        totalGst:
          type: number
        financialYear:
          type: string
    GstReportEntry:
      type: object
      required: [month, monthName, cgst, sgst, igst, totalGst]
      properties:
        month:
          type: integer
        monthName:
          type: string
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        totalGst:
          type: number
        taxableAmount:
          type: number
    ProfitReport:
      type: object
      required: [entries, totalRevenue, totalCost, totalProfit]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ProfitReportEntry"
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        profitMargin:
          type: number
    ProfitReportEntry:
      type: object
      required: [period, revenue, cost, profit]
      properties:
        period:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        margin:
          type: number
    ProductWiseReport:
      type: object
      required: [productId, productName, totalQuantity, totalRevenue, totalProfit]
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        totalGst:
          type: number
    CustomerWiseReport:
      type: object
      required: [customerId, customerName, totalPurchases, invoiceCount]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        totalPurchases:
          type: number
        invoiceCount:
          type: integer
        lastPurchaseDate:
          type: string
...
    DeleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string