# lib/api-zod/src/generated/types/customerWiseReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerWiseReport {
  customerId: number;
  customerName: string;
  totalPurchases: number;
  invoiceCount: number;
  lastPurchaseDate?: string;
}