# lib/api-zod/src/generated/types/salesReportEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */

export interface SalesReportEntry {
  period: string;
  sales: number;
  gst: number;
  profit?: number;
  invoiceCount: number;
}